requirements.txt
asgiref==3.3.4
pytz==2021.1
sqlparse==0.4.1
Pillow
social_auth_app_django
django-environ